Status widget improvements
