+++ frontend/src/components/core/StatusWidget/styled-components.ts
@@ -51,6 +51,7 @@ export const StyledConnectionStatusLabel = styled.label<
     "opacity 500ms 0ms, clip 500ms 0ms, max-width 500ms 0ms, margin 500ms 0ms, visibility 0ms 500ms",
   opacity: isMinimized ? 0 : 1,
   visibility: isMinimized ? "hidden" : "visible",
+  lineHeight: 1,
 }))
 
 /*
