+++ frontend/src/components/shared/Tooltip/styled-components.tsx
@@ -16,6 +16,7 @@
  */
 
 import styled from "@emotion/styled"
+import { transparentize } from "color2k"
 
 export const StyledWrapper = styled.div({
   display: "table",
@@ -29,3 +30,22 @@ export const StyledEllipsizedDiv = styled.div({
   whiteSpace: "nowrap",
   display: "table-cell",
 })
+
+export const StyledTooltipContentWrapper = styled.div(({ theme }) => ({
+  boxSizing: "border-box",
+  fontSize: `${theme.fontSizes.sm} !important`,
+  maxWidth: `calc(${theme.sizes.contentMaxWidth} - 4rem)`,
+  maxHeight: "300px",
+  overflow: "auto",
+  padding: `${theme.spacing.xs} ${theme.spacing.md}`,
+
+  [`@media (max-width: ${theme.breakpoints.sm})`]: {
+    maxWidth: `calc(100% - 2rem)`,
+  },
+  img: {
+    maxWidth: "100%",
+  },
+  code: {
+    background: transparentize(theme.colors.darkenedBgMix60, 0.8),
+  },
+}))
