+++ frontend/src/components/shared/TooltipIcon/index.tsx
@@ -1,3 +1,2 @@
 export * from "./TooltipIcon"
 export { default } from "./TooltipIcon"
-export { StyledTooltipContentWrapper } from "./styled-components"
