+++ frontend/src/components/shared/TooltipIcon/styled-components.ts
@@ -1,5 +1,4 @@
 import styled from "@emotion/styled"
-import { transparentize } from "color2k"
 
 export const StyledTooltipIconWrapper = styled.div(({ theme }) => ({
   display: "flex",
@@ -13,24 +12,5 @@ export const StyledTooltipIconWrapper = styled.div(({ theme }) => ({
     svg: {
       stroke: theme.colors.fadedText60,
     },
-  },
-}))
-
-export const StyledTooltipContentWrapper = styled.div(({ theme }) => ({
-  boxSizing: "border-box",
-  fontSize: `${theme.fontSizes.sm} !important`,
-  maxWidth: `calc(${theme.sizes.contentMaxWidth} - 4rem)`,
-  maxHeight: "300px",
-  overflow: "auto",
-  padding: `${theme.spacing.xs} ${theme.spacing.md}`,
-
-  [`@media (max-width: ${theme.breakpoints.sm})`]: {
-    maxWidth: `calc(100% - 2rem)`,
-  },
-  img: {
-    maxWidth: "100%",
-  },
-  code: {
-    background: transparentize(theme.colors.darkenedBgMix60, 0.8),
   },
 }))
